adds vstack and associated tests. HStack sets env appropriately to get the right divider/spacer behavior

+++ src/font.rs
@@ -1,19 +1,17 @@
-use crate::primitives::uint;
-
 pub trait Font {
     /// The height of a character in points
-    fn line_height(&self) -> uint;
+    fn line_height(&self) -> u16;
     /// The width of a character in points
-    fn character_width(&self, character: char) -> uint;
+    fn character_width(&self, character: char) -> u16;
 }
 
 #[derive(Default)]
 pub struct TextBufferFont();
 impl Font for TextBufferFont {
-    fn line_height(&self) -> uint {
+    fn line_height(&self) -> u16 {
         1
     }
-    fn character_width(&self, _character: char) -> uint {
+    fn character_width(&self, _character: char) -> u16 {
         1
     }
 }
